fix permissions in makefile, update grep / sed in rawk
makefile was setting the share directory to -R 755 but files such
as rawkrc should not have +x. files are now set to 644 and dirs to
755.

greps in rawk are now grep -ve to explictly set regex.

the stylesheet sed in rawk was sed -s REGEX - this is valid in linux
so despite being a typo, the command would work in linux but not
openbsd. this has been fixed to -e.