Fix an apparent typo in the default CSS style.
Due to a capitalisation mismatch between the style and the header
template, the "headerSubTitle"/"headerSubtitle" style was never used.

Re-matching the capitalisation (I made the admittedly arbitrary
choice in favor of the lowercase "t".), and thereby causing the style
to actually be used, produced what I thought was a somewhat jarring
typographical effect in the header. This was solved by commenting-out
the font-size setting from the headerSubtitle style. Hopefully, the
subtitle is visible yet clearly subordinate to, and clearly separate
from, the title.